backend/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.2.17
django-bootstrap5==24.2
django-cors-headers==4.4.0
django-filter==21.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.7.0
exceptiongroup==1.2.2
Faker==12.0.1
filetype==1.2.0
flake8==5.0.4
flake8-docstrings==1.7.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.7.0
mixer==7.2.2
more-itertools==10.5.0
oauthlib==3.2.0
packaging==21.3
pep8-naming==0.13.3
Pillow==9.0.1
pluggy==0.13.1
psycopg2==2.9.9
psycopg2-binary==2.9.3 
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==7.4.4
pytest-django==4.9.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
pytest-subtests==0.13.1
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2021.3
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.5
sqlparse==0.4.2
types-requests==2.32.0.20241016
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.7.0
